6. Идеальный вес.[#250303]

+++ src/ru/job4j/converter/Converter.java
@@ -6,8 +6,8 @@ public class Converter {
         return rsl;
     }
 
-    public static float rubleToDollar(float value) {
-        float rsl = value /60;
+    public static int rubleToDollar(int value) {
+        int rsl = value /60;
         return rsl;
     }
